add minishell/cub3d and gnl extras

+++ ft_bzero.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vminomiy <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: vminomiy <vminomiy@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/01/21 17:32:35 by vminomiy          #+#    #+#             */
-/*   Updated: 2020/01/27 16:43:18 by vminomiy         ###   ########.fr       */
+/*   Created: 2020/11/27 16:59:29 by vminomiy          #+#    #+#             */
+/*   Updated: 2020/11/27 16:59:29 by vminomiy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned long int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
